Item 5043 Compute accession register when elimnation action occurs. Correct bug 5146 et 5157

See merge request vitam/vitam!4921

+++ sources/worker/worker-core/src/main/java/fr/gouv/vitam/worker/core/plugin/dip/ManifestBuilder.java
@@ -217,7 +217,7 @@ public class ManifestBuilder implements AutoCloseable {
                         String[] dataObjectVersion = minimalDataObjectType.getDataObjectVersion().split("_");
 
                         maps.put(minimalDataObjectType.getId(),
-                            AccessLogUtils.getWorkerInfo(dataObjectVersion[0], Integer.parseInt(dataObjectVersion[1]), binaryDataObjectType.getSize().intValue(), linkedAU, fileName));
+                            AccessLogUtils.getWorkerInfo(dataObjectVersion[0], Integer.parseInt(dataObjectVersion[1]), binaryDataObjectType.getSize().longValue(), linkedAU, fileName));
                     }
                     marshaller.marshal(minimalDataObjectType, writer);
                 }
